requirements_email_intake.txt
# Gmail OAuth and Email Processing
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.111.0

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3

# Email parsing and utilities
email-validator==2.1.0
python-dateutil==2.8.2